add validations

+++ controller/users.js
@@ -5,11 +5,25 @@ var passwordHash = require('../helpers/hashPassword')
 
 
 
-var authorizedRole = (req,res,next)=> {
-  if(req.headers.token){
-    next()
+var authorizedRole= (req,res,next)=> {
+  let token =req.headers.token
+  if(token){
+    jwt.verify(token, 'rahasiacoy', function(err, decoded) {
+    console.log(decoded)
+      if(decoded.role==='admin'){
+      console.log(decoded.role);
+        req.body.role='admin'
+        next()
+      }else if(decoded.role==='user'){
+      console.log(decoded.role);
+        req.body.role='user'
+        next()
+      }
+      else res.send({message:'login Admin cyin'})
+
+    });
   }else {
-    res.send({message:'error'})
+    res.send({message:'login cyin'})
   }
 }
 
@@ -20,8 +34,11 @@ var signup = (req,res) => {
     db.User.create({name:req.body.name,username:req.body.username,password:password,email:req.body.email,role:req.body.role})
     .then(user => {
       res.send(user)
-    })
-  })
+    }).
+    catch(err => {
+     res.send(err.message)
+     })
+   })
 }
 
 var login = (req,res) => {
@@ -43,10 +60,14 @@ var login = (req,res) => {
 }
 
 var getAllUser = (req,res) => {
-  db.User.findAll()
-  .then(users => {
-    res.send(users)
-  })
+  if(req.body.role==='admin'){
+    db.User.findAll()
+    .then(users => {
+      res.send(users)
+    })
+  }else {
+    res.send('Login Admin cyin')
+  }
 }
 
 var getUserById = (req,res) => {
@@ -58,11 +79,19 @@ var getUserById = (req,res) => {
 }
 
 var createUser = (req,res) => {
-
-  db.User.create(req.body).
-  then(user => {
-    res.send(user)
-  })
+if(req.body.role==='admin'){
+    db.User.create(req.body).
+    then(user => {
+      res.send(user)
+    }).
+    catch(err =>{
+      console.log(err);
+      res.send('err coy')
+    })
+  }
+  else {
+    res.send('Login Admin Cyin')
+  }
 }
 
 var updateUser = (req,res) => {
@@ -74,10 +103,14 @@ var updateUser = (req,res) => {
 }
 
 var deleteUser = (req,res) => {
-  db.User.destroy({where:{id:req.id}})
-  .then((user) => {
-    res.send(`${user.name} has been deleted`)
-  })
+  if(req.body.role==='admin'){
+    db.User.destroy({where:{id:req.id}})
+    .then((user) => {
+      res.send(`${user.name} has been deleted`)
+    })
+  }else {
+    res.send('Login Admin Cyin')
+  }
 }
 
 module.exports = {
